+++ sdk/eventhub/Azure.Messaging.EventHubs/samples/Sample01_HelloWorld.cs
@@ -5,7 +5,7 @@ using System;
 using System.Threading.Tasks;
 using Azure.Messaging.EventHubs.Metadata;
 using Azure.Messaging.EventHubs.Samples.Infrastructure;
-using Azure.Messaging.EventHubs.Samples.Infrastructure.Interfaces;
+using Azure.Messaging.EventHubs.Samples.Infrastructure.Means;
 
 namespace Azure.Messaging.EventHubs.Samples
 {
