+++ sdk/eventhub/Azure.Messaging.EventHubs/samples/Sample12_ConsumeEventsWithEventProcessor.cs
@@ -9,7 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Azure.Messaging.EventHubs.Processor;
 using Azure.Messaging.EventHubs.Samples.Infrastructure;
-using Azure.Messaging.EventHubs.Samples.Infrastructure.Interfaces;
+using Azure.Messaging.EventHubs.Samples.Infrastructure.Means;
 
 namespace Azure.Messaging.EventHubs.Samples
 {
